pyproject.toml
[tool.poetry]
name = "bankly"
version = "0.1.0"
description = ""
authors = ["mustafausama <mustafausama@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
